Backend/requirements.txt
django==4.2.11
djangorestframework==3.14.0
markdown==3.5.2
django-filter==24.1
django-cors-headers
mysqlclient==2.2.4
